fix(tilecatalog): rename old and new tilecatalog
