+++ src/components/TileCatalog/tileCatalogReducer.js
