+++ src/components/TileCatalog/tileCatalogReducer.test.js
