+++ src/styles.scss
@@ -193,5 +193,5 @@ $deprecations--message: 'Deprecated code was found, this code will be removed be
 @import 'components/List/list';
 @import 'components/List/ListItem/list-item';
 @import 'components/List/ListHeader/list-header';
-@import 'components/TileCatalog/tile-catalog';
+@import 'components/TileCatalogNew/tile-catalog';
 @import 'components/ValueCard/value-card';
